[add#UnixTools] Init new project `wc` on `C`
Add "Makefile" for build this project.

wc [OPTION]... [FILE]...

"wc" - unix tools, which print newline, word, and byte cout for each
file. Print newline, word, and byte counts for each FILE, and a total
line if more than one FILE is specified. With no FILE, or when FILE is
-, read standart input.

Options:

    -c, --bytes
        print the byte counts

    -w, --words
        print the word counts

    -l, --lines
        print the newline counts

    -L, --max-line-length
        print the maximum display width