[feat#UnixTools/wc] Add Correct parsing of program operation flags
Added new states for "Mode": not initialized. Also added function of
flag initialization.

Resolved a problem that occurred when no flags were specified and
the application mode was "uninitialized".
Thanks to the "Mode" flag NO_INIT this can be detected and the
default mode - PRI_LWB - can be set by condition.